+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [0.4.0] - 2024-03-03
 ### Added
 - JavaScript evaluation raises `JsError` for uncaught JavaScript exceptions. Evalution includes `JsFunction` calls, and `Runtime` methods `eval`, `mod_evaluate`, `call`.
     - For function calls, the raised `JsError` object stores the thrown JavaScript exception in the attribute `value`.
@@ -27,7 +29,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.1.0] - 2024-01-19
 Initial release: module loading, code evaluation, type conversions, function calls.
 
-[Unreleased]: https://github.com/ubolonton/denopy/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/ubolonton/denopy/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/ubolonton/denopy/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/ubolonton/denopy/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/ubolonton/denopy/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ubolonton/denopy/compare/6d975ef1...v0.1.0